+++ shared/messaging/utils_test.go
@@ -1,0 +1,21 @@
+package messaging
+
+import (
+	"testing"
+
+	"github.com/golang/protobuf/proto"
+	"github.com/stretchr/testify/require"
+)
+
+func Test_UnpackMessage_roundtrip(t *testing.T) {
+	someMsg := &DebugMessage{Action: DebugMessageAction_DEBUG_PERSISTENCE_CLEAR_STATE}
+	packedMsg, err := PackMessage(someMsg)
+	require.NoError(t, err)
+
+	unpackedMsg, err := UnpackMessage(packedMsg)
+	require.NoError(t, err)
+
+	if !proto.Equal(someMsg, unpackedMsg.(proto.Message)) {
+		t.Fail()
+	}
+}